Adapt to Raspberry Pi OS Bullseye
Short explanation:
With the default instructions, the input_raspicam.so would no be
generated.
In the new RPi OS, raspistill raspivid have been replaced by libcamera.
The CMakeLists.txt file would check for /opt/vc, which does not exist in
the new RPi OS. This files, instead, are in /usr.

To make this library work, it is also necessary to comment "camera_auto_detect=1",
and add "start_x=1" and "gpu_mem=128" in /boot/config.txt.
After rebooting, this library was working.